requirements.txt
Deprecated~=1.2.14
Faker~=22.0.0
GitPython~=3.1.40
Jinja2~=3.1.2
Markdown~=3.5.1
MarkupSafe~=2.1.3
Pillow~=10.1.0
PyYAML~=6.0.1
altair~=5.2.0
attrs~=23.2.0
beautifulsoup4~=4.12.2
blinker~=1.7.0
cachetools~=5.3.2
certifi~=2023.11.17
charset-normalizer~=3.3.2
click~=8.1.7
colorama~=0.4.6
contourpy~=1.2.0
cycler~=0.12.1
entrypoints~=0.4
favicon~=0.7.0
fonttools~=4.47.0
gensim~=4.3.2
gitdb~=4.0.11
htbuilder~=0.6.2
idna~=3.6
importlib-metadata~=6.11.0
jieba~=0.42.1
jsonschema~=4.20.0
jsonschema-specifications~=2023.12.1
kiwisolver~=1.4.5
lxml~=5.0.0
markdown-it-py~=3.0.0
markdownlit~=0.0.7
matplotlib~=3.8.2
mdurl~=0.1.2
memory-profiler~=0.61.0
more-itertools~=10.1.0
numpy~=1.26.2
packaging~=23.2
pandas~=2.1.4
pip~=23.3.2
prettytable~=3.9.0
protobuf~=4.25.1
psutil~=5.9.7
pyarrow~=14.0.2
pydeck~=0.8.1b0
pyecharts~=2.0.4
pygments~=2.17.2
pymdown-extensions~=10.7
pyparsing~=3.1.1
python-dateutil~=2.8.2
pytz~=2023.3.post1
referencing~=0.32.0
requests~=2.31.0
rich~=13.7.0
rpds-py~=0.16.2
scipy~=1.11.4
setuptools~=69.0.3
simplejson~=3.19.2
six~=1.16.0
smart-open~=6.4.0
smmap~=5.0.1
soupsieve~=2.5
st-annotated-text~=4.0.1
streamlit~=1.29.0
streamlit-camera-input-live~=0.2.0
streamlit-card~=1.0.0
streamlit-carousel~=0.0.4
streamlit-echarts~=0.4.0
streamlit-embedcode~=0.1.2
streamlit-extras~=0.3.6
streamlit-faker~=0.0.3
streamlit-image-coordinates~=0.1.6
streamlit-keyup~=0.2.2
streamlit-toggle-switch~=1.0.2
streamlit-vertical-slider~=2.5.5
tenacity~=8.2.3
toml~=0.10.2
toolz~=0.12.0
tornado~=6.4
tqdm~=4.65.2
typing-extensions~=4.9.0
tzdata~=2023.4
tzlocal~=5.2
urllib3~=2.1.0
validators~=0.22.0
watchdog~=3.0.0
wcwidth~=0.2.12
wheel~=0.41.2
wrapt~=1.16.0
zipp~=3.17.0